Separate socket2 into new file.
